refactor: Remove TournamentCard component and update imports
- Deleted the TournamentCard component to streamline tournament display logic.
- Updated imports in TournamentForm and TournamentList to reference the new TournamentView component for consistency.
